feat: Add support for deleting units by custom ID
This commit updates the `deleteUnit` function in the `unit.controller.js` file to allow deleting units by custom ID in addition to the default ID. It checks if the provided ID is a valid Object ID and deletes the unit accordingly. If the ID is not a valid Object ID, it deletes the unit by the custom ID field. This enhancement improves the flexibility of unit deletion and allows for more efficient deletions.

Co-authored-by: [Goodluck Japhet] <[goodluckjaphet7@gmail.com]>