refactor: Update unit.route.js to use deleteMultipleUnits and hardDeleteUnits
This commit updates the `unit.route.js` file to use the `deleteMultipleUnits` and `hardDeleteUnits` functions from the `unit.controller.js` file. The `deleteAllUnits` and `deleteUnitsWithRef` functions have been replaced with these new functions to improve the clarity and consistency of the codebase. This refactor aligns with the recent changes made to the unit deletion functionality.

Co-authored-by: [Goodluck Japhet] <[goodluckjaphet7@gmail.com]>