feat: Add axios dependency and Admin model to support admin authentication
This commit adds the axios dependency to the package.json file and the Admin model to the api/prisma/schema.prisma file. The axios dependency is required for making HTTP requests, and the Admin model defines the structure of the admin user in the database. These changes are necessary to support admin authentication in the application.

Co-authored-by: [Goodluck Japhet] <[goodluckjaphet7@gmail.com]>