feat: Update login functionality to store token in local storage
This commit updates the login functionality in the `authService.js` file and the `Login.vue` component to store the token in local storage upon successful login. It removes the unnecessary check for the presence of the token in the response data and directly returns the full response object. The `handleLogin` function in the `Login.vue` component now redirects to the dashboard route after successful login. Additionally, an error message is displayed if the login fails or no token is received.

Co-authored-by: [Goodluck Japhet] <[goodluckjaphet7@gmail.com]>