feat: Update axios and dotenv dependencies
This commit updates the axios and dotenv dependencies in the package.json and package-lock.json files. The axios dependency is updated to version 1.7.3, and the dotenv dependency is added with version 16.4.5. These updates ensure that the application uses the latest stable versions of these libraries, which may improve performance, security, and compatibility.

Co-authored-by: [Goodluck Japhet] <[goodluckjaphet7@gmail.com]>