feat: Update axios and dotenv dependencies
This commit updates the axios and dotenv dependencies to their latest versions. Axios is updated from 1.7.3 to 1.7.4, and dotenv is added with version 16.4.5. These updates ensure that the application uses the latest stable versions of these libraries, which may improve performance, security, and compatibility.

removes svg and also made blue theme as primary